# Improvement - Performance of the current VTK writer in t8code