**_Describe your changes here:_**
The current version of the t8_vtk_writer uses the native classes for VTK cell types to add cells. Instantiating these classes do a lot of memory allocations not required here. We can directly use vtkIdType arrays for storing the point ids for a cell and just insert these to the vtkCellArray. This avoids allocation overheads using the native types. Additionally, we 
can already allocate memory for the points, cells, and fields earlier and avoid dynamic resizing of these arrays.

Additionally, when inserting points to the point array, duplications can be avoided by using the vtkMergePoints data structure.

Performance measurements:

| Elements (Hexa)  | Points     | Without Opt. | With Alloc  | With vtkIdType Arrays  | Speedup | 
|------------------|------------|--------------|-------------| -----------------------|---------| 
| 1.519.449        | 12.155.592 | 17646.3ms    | 14352.3ms   | 1045.2ms               | 16.8x   |  

Another feature is to use the vtkMergePoints data structure in order to avoid writing out duplicated points for the shared vertices of elements. 

| Elements (Hexa)  | Points Before | With vtkMergePoints | Runtime | 
|------------------|---------------|---------------------| ------- | 
| 1.519.449        | 12.155.592    | 2.423.685           | 3742.4ms|

**_All these boxes must be checked by the AUTHOR before requesting review:_**
- [ ] The PR is *small enough* to be reviewed easily. If not, consider splitting up the changes in multiple PRs.
- [ ] The title starts with one of the following prefixes: `Documentation:`, `Bugfix:`, `Feature:`, `Improvement:` or `Other:`.
- [ ] If the PR is related to an issue, make sure to link it.
- [ ] The author made sure that, as a reviewer, he/she would check all boxes below.

**_All these boxes must be checked by the REVIEWERS before merging the pull request:_**

As a reviewer please read through all the code lines and make sure that the code is *fully understood, bug free, well-documented and well-structured*.
#### General
- [ ] The reviewer executed the new code features at least once and checked the results manually.
- [ ] The code follows the [t8code coding guidelines](https://github.com/DLR-AMR/t8code/wiki/Coding-Guideline).
- [ ] New source/header files are properly added to the CMake files.
- [ ] The code is well documented. In particular, all function declarations, structs/classes and their members have a proper doxygen documentation.
- [ ] All new algorithms and data structures are sufficiently optimal in terms of memory and runtime (If this should be merged, but there is still potential for optimization, create a new issue).
#### Tests
- [ ] The code is covered in an existing or new test case using Google Test.
- [ ] Valgrind doesn't find any bugs in the new code. [This script](https://github.com/DLR-AMR/t8code/blob/main/scripts/check_valgrind.sh) can be used to check for errors; see also this [wiki article](https://github.com/DLR-AMR/t8code/wiki/Debugging-with-valgrind).

If the Pull request introduces code that is not covered by the github action (for example coupling with a new library):
  - [ ] Should this use case be added to the github action?
  - [ ] If not, does the specific use case compile and all tests pass (check manually).
#### Scripts and Wiki
- [ ] If a new directory with source files is added, it must be covered by the `script/find_all_source_files.scp` to check the indentation of these files.
- [ ] If this PR introduces a new feature, it must be covered in an example or tutorial and a Wiki article.
#### License
- [ ] The author added a BSD statement to `doc/` (or already has one).