@Davknapp 2025-05-16
We need to compute the local bounds of a cmesh to pass the bounds to vtkMergePoints via something like:
```
    points->InitPointInsertion (points_store, unstructuredGrid->GetBounds ());
```