@Davknapp 2025-05-16
> We need to compute the local bounds of a cmesh to pass the bounds to vtkMergePoints via something like:
> 
> ```
>     points->InitPointInsertion (points_store, unstructuredGrid->GetBounds ());
> ```

@Flatken you can try to merge with #1659 to use a first version of this computation. This is not reviewed yet and might change, though. The interface will probably stay the same.